Libs: Move SI string conversion to units
Move si_str_to_float() and float_to_si_string() from libs/util to
libs/units
